requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
browserforge==1.2.3
cachetools==6.2.1
camoufox==0.4.11
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Faker==37.12.0
frozenlist==1.8.0
geoip2==5.1.0
google-api-core==2.28.0
google-api-python-client==2.185.0
google-auth==2.42.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
httplib2==0.31.0
idna==3.11
imagekitio==4.2.0
language-tags==1.2.0
lxml==6.0.2
maxminddb==2.8.2
multidict==6.7.0
numpy==2.3.4
oauthlib==3.3.1
orjson==3.11.4
packaging==25.0
platformdirs==4.5.0
playwright==1.55.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
pyparsing==3.2.5
PySocks==1.7.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
screeninfo==0.8.1
soupsieve==2.8
telebot==0.0.5
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
